# openapi.yaml
openapi: 3.0.3
info:
  title: Fath (فتح) | Saudi Open Banking Sandbox API
  description: |
    Saudi Open Banking Sandbox for Developers | بيئة اختبار البنك المفتوح للمطورين

    **⚠️ Regulatory Disclaimer:**
    Fath is an independent portfolio project for educational purposes. It is not affiliated with, endorsed by, or certified by SAMA (Saudi Central Bank). All bank names and financial data are fictional and generated for testing purposes only.

    Reference standard: SAMA Open Banking Framework Technical Standards v1.0
  version: 1.0.0
  contact:
    name: Yamen Alwaily
    url: https://github.com/yamen-alwaily/fath
servers:
  - url: http://localhost:5000
    description: Local Sandbox Development Server
paths:
  /oauth/token:
    post:
      summary: OAuth 2.0 Token Exchange with PKCE
      description: Exchange an authorization code + PKCE code_verifier for a Bearer access token.
      tags:
        - Security & OAuth 2.0
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - grant_type
                - client_id
                - code
              properties:
                grant_type:
                  type: string
                  example: authorization_code
                client_id:
                  type: string
                  example: sandbox_client_demo
                code:
                  type: string
                  example: auth_code_12345
                code_verifier:
                  type: string
                  example: dBjftJeZ4CVP-mB92K27uhbUJU1p1r_wW1gFWFOEjXk
      responses:
        '200':
          description: Token successfully issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Invalid authorization code or PKCE verification failure
  /oauth/revoke:
    post:
      summary: Revoke Token / Consent Grant
      description: Revoke an active access token or consent grant.
      tags:
        - Security & OAuth 2.0
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Revocation status
  /oauth/introspect:
    get:
      summary: Introspect Token (RFC 7662)
      tags:
        - Security & OAuth 2.0
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Active status and scope claims of the token
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Bearer Token
  schemas:
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          example: 3600
        scope:
          type: string
          example: accounts:read balances:read transactions:read
...